+++ ml/layers/kan/discrete.go
@@ -385,7 +385,7 @@ func (c *Config) scheduledSoftness(scope *model.Scope, base *Node) *Node {
 
 	// Calculate scheduleTime: from 0.0 to 1.0 depending on training progress.
 	globalStep := ConvertDType(optimizer.GetGlobalStepVar(rootScope).NodeValue(g), dtypes.Float32)
-	lastStep := ConvertDType(train.GetTrainLastStepVar(rootScope).NodeValue(g), dtypes.Float32)
+	lastStep := ConvertDType(train.GetTrainLastStepVar(g).NodeValue(g), dtypes.Float32)
 	// scheduleTime will be at most 1.0, if for some reason globalStep >
 	scheduleTime := MinScalar(Div(globalStep, MaxScalar(lastStep, 1.0)), 1.0)
 	zero := ZerosLike(lastStep)