+++ ml/layers/layers.go
@@ -437,7 +437,7 @@ func AddL2Regularization(scope *model.Scope, amount *Node, values ...*Node) {
 		}
 	}
 	loss = Mul(loss, amount)
-	train.AddLoss(scope, loss)
+	train.AddLoss(loss)
 }
 
 // Normalize shifts and scales the input such that the mean becomes zero and the variance one.